pyproject.toml
[project]
name = "registro-futbol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "django>=5.2.4",
  "django-filter>=25.1",
  "djangorestframework>=3.16.0",
  "markdown>=3.8.2",
  "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.1",
    "djangorestframework-stubs>=3.16.0",
]
